requirements.txt
Django==2.0.7
django-ckeditor==5.5.0
django-js-asset==1.1.0
Pillow==5.2.0
pytz==2018.5
